# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  handle: string;
  avatarUrl: string;
  joinedAt: string;
  bio: string;
  location: string;
}

export type RequestSummaryStatus =
  (typeof RequestSummaryStatus)[keyof typeof RequestSummaryStatus];

export const RequestSummaryStatus = {
  open: "open",
  fulfilled: "fulfilled",
  closed: "closed",
} as const;

export type RequestSummaryUrgency =
  (typeof RequestSummaryUrgency)[keyof typeof RequestSummaryUrgency];

export const RequestSummaryUrgency = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface RequestSummary {
  id: string;
  title: string;
  description: string;
  category: string;
  budgetMin?: number | null;
  budgetMax?: number | null;
  status: RequestSummaryStatus;
  urgency: RequestSummaryUrgency;
  location: string;
  createdAt: string;
  buyer: User;
  responseCount: number;
}

export interface UserProfile {
  user: User;
  requestCount: number;
  responseCount: number;
  acceptedCount: number;
  recentRequests?: RequestSummary[];
}

export interface UpdateUserBody {
  name?: string;
  bio?: string;
  location?: string;
  avatarUrl?: string;
}

export type SellerResponseCondition =
  (typeof SellerResponseCondition)[keyof typeof SellerResponseCondition];

export const SellerResponseCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export type SellerResponseStatus =
  (typeof SellerResponseStatus)[keyof typeof SellerResponseStatus];

export const SellerResponseStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export interface SellerResponse {
  id: string;
  requestId: string;
  seller: User;
  price: number;
  condition: SellerResponseCondition;
  message: string;
  photos: string[];
  status: SellerResponseStatus;
  threadId?: string | null;
  createdAt: string;
}

export type RequestDetail = RequestSummary & {
  tags?: string[];
  responses?: SellerResponse[];
};

export type CreateRequestBodyUrgency =
  (typeof CreateRequestBodyUrgency)[keyof typeof CreateRequestBodyUrgency];

export const CreateRequestBodyUrgency = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface CreateRequestBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: string;
  budgetMin?: number;
  budgetMax?: number;
  location?: string;
  urgency?: CreateRequestBodyUrgency;
  tags?: string[];
}

export type UpdateRequestBodyStatus =
  (typeof UpdateRequestBodyStatus)[keyof typeof UpdateRequestBodyStatus];

export const UpdateRequestBodyStatus = {
  open: "open",
  fulfilled: "fulfilled",
  closed: "closed",
} as const;

export type UpdateRequestBodyUrgency =
  (typeof UpdateRequestBodyUrgency)[keyof typeof UpdateRequestBodyUrgency];

export const UpdateRequestBodyUrgency = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface UpdateRequestBody {
  title?: string;
  description?: string;
  status?: UpdateRequestBodyStatus;
  urgency?: UpdateRequestBodyUrgency;
}

export type CreateResponseBodyCondition =
  (typeof CreateResponseBodyCondition)[keyof typeof CreateResponseBodyCondition];

export const CreateResponseBodyCondition = {
  new: "new",
  like_new: "like_new",
  good: "good",
  fair: "fair",
  used: "used",
} as const;

export interface CreateResponseBody {
  price: number;
  condition: CreateResponseBodyCondition;
  /** @minLength 1 */
  message: string;
  photos: string[];
}

export interface Thread {
  id: string;
  requestId: string;
  requestTitle: string;
  otherUser: User;
  lastMessage: string;
  updatedAt: string;
  unread: number;
}

export interface Message {
  id: string;
  threadId: string;
  sender: User;
  body: string;
  createdAt: string;
}

export interface ThreadDetail {
  id: string;
  request: RequestSummary;
  participants: User[];
  messages: Message[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface OverviewStats {
  openRequests: number;
  totalResponses: number;
  fulfilledThisWeek: number;
  activeBuyers: number;
  activeSellers: number;
}

export type ActivityEventKind =
  (typeof ActivityEventKind)[keyof typeof ActivityEventKind];

export const ActivityEventKind = {
  request_created: "request_created",
  response_created: "response_created",
  response_accepted: "response_accepted",
  request_fulfilled: "request_fulfilled",
} as const;

export interface ActivityEvent {
  id: string;
  kind: ActivityEventKind;
  createdAt: string;
  actor: User;
  summary: string;
  requestId?: string | null;
}

export type SwitchUserBody = {
  userId: string;
};

export type ListRequestsParams = {
  search?: string;
  category?: string;
  status?: ListRequestsStatus;
  buyerId?: string;
  sort?: ListRequestsSort;
};

export type ListRequestsStatus =
  (typeof ListRequestsStatus)[keyof typeof ListRequestsStatus];

export const ListRequestsStatus = {
  open: "open",
  fulfilled: "fulfilled",
  closed: "closed",
} as const;

export type ListRequestsSort =
  (typeof ListRequestsSort)[keyof typeof ListRequestsSort];

export const ListRequestsSort = {
  newest: "newest",
  oldest: "oldest",
  most_responses: "most_responses",
  urgent: "urgent",
} as const;

export type UpdateResponseStatusBodyStatus =
  (typeof UpdateResponseStatusBodyStatus)[keyof typeof UpdateResponseStatusBodyStatus];

export const UpdateResponseStatusBodyStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
} as const;

export type UpdateResponseStatusBody = {
  status: UpdateResponseStatusBodyStatus;
};

export type SendMessageBody = {
  /** @minLength 1 */
  body: string;
};