# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Desperately Seeking - buyer-first marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List buyer requests
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [open, fulfilled, closed]
        - in: query
          name: buyerId
          schema:
            type: string
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest, most_responses, urgent]
            default: newest
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestSummary"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a buyer request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
  /requests/{requestId}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get full detail of a buyer request including responses
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update status / details of a request (buyer only)
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
...
  /threads/{threadId}:
    get:
      operationId: getThread
      tags: [messages]
      summary: Get thread with messages
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Thread detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
  /threads/{threadId}/messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a thread
      parameters:
        - in: path
          name: threadId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, handle, avatarUrl, joinedAt, bio, location]
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: string
        avatarUrl:
          type: string
        joinedAt:
          type: string
          format: date-time
        bio:
          type: string
        location:
          type: string
...
    RequestSummary:
      type: object
      required:
        [
          id,
          title,
          description,
          category,
          status,
          urgency,
          createdAt,
          buyer,
          responseCount,
          location,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        budgetMin:
          type: number
          nullable: true
        budgetMax:
          type: number
          nullable: true
        status:
          type: string
          enum: [open, fulfilled, closed]
        urgency:
          type: string
          enum: [low, normal, high]
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        buyer:
          $ref: "#/components/schemas/User"
        responseCount:
          type: integer
    RequestDetail:
      allOf:
        - $ref: "#/components/schemas/RequestSummary"
        - type: object
          properties:
            tags:
              type: array
              items:
                type: string
            responses:
              type: array
              items:
                $ref: "#/components/schemas/SellerResponse"
    CreateRequestBody:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        category:
          type: string
        budgetMin:
          type: number
        budgetMax:
          type: number
        location:
          type: string
        urgency:
          type: string
          enum: [low, normal, high]
        tags:
          type: array
          items:
            type: string
    UpdateRequestBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, fulfilled, closed]
        urgency:
          type: string
          enum: [low, normal, high]
    SellerResponse:
      type: object
      required:
        [
          id,
          requestId,
          seller,
          price,
          condition,
          message,
          photos,
          status,
          createdAt,
        ]
      properties:
        id:
          type: string
        requestId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        price:
          type: number
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, accepted, declined]
        threadId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    ThreadDetail:
      type: object
      required: [id, request, participants, messages]
      properties:
        id:
          type: string
        request:
          $ref: "#/components/schemas/RequestSummary"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/User"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    Message:
      type: object
      required: [id, threadId, sender, body, createdAt]
      properties:
        id:
          type: string
        threadId:
          type: string
        sender:
          $ref: "#/components/schemas/User"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
...