# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRequestBodyUrgency } from "./createRequestBodyUrgency";

export interface CreateRequestBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  category: string;
  budgetMin?: number;
  budgetMax?: number;
  location?: string;
  urgency?: CreateRequestBodyUrgency;
  tags?: string[];
}