# lib/api-zod/src/generated/types/requestDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { RequestSummary } from "./requestSummary";
import type { SellerResponse } from "./sellerResponse";

export type RequestDetail = RequestSummary & {
  tags?: string[];
  responses?: SellerResponse[];
};