# lib/api-zod/src/generated/types/requestSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { RequestSummaryStatus } from "./requestSummaryStatus";
import type { RequestSummaryUrgency } from "./requestSummaryUrgency";
import type { User } from "./user";

export interface RequestSummary {
  id: string;
  title: string;
  description: string;
  category: string;
  budgetMin?: number | null;
  budgetMax?: number | null;
  status: RequestSummaryStatus;
  urgency: RequestSummaryUrgency;
  location: string;
  createdAt: Date;
  buyer: User;
  responseCount: number;
}