# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  handle: string;
  avatarUrl: string;
  joinedAt: Date;
  bio: string;
  location: string;
}